requirements.txt
scipy
altgraph==0.17.4
macholib==1.16.3
numpy==2.3.4
packaging==25.0
pillow==12.0.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
setuptools==80.9.0
opencv-python>=4.5.0

